Rename new utils file
